+++ omar qt/global.cpp
@@ -12,6 +12,7 @@ int kit_kat_index = 0;
 int switching_index = 0;
 int arr_sub_insights[1000] = { 0 };
 int admin_chosen_user;
+int chosenSubscriptionIndex = -1;
 // Short integers
 short indexofuser = 0;
 short number_of_users_in_array = 0;
